bot.py
import discord
from discord.ext import commands
import aiohttp


description = """Work in progress."""

bot = commands.Bot(command_prefix='r!', description=description)


@bot.event
async def on_ready():
    print("Logged in")


@bot.event
async def on_message(message):
    await bot.process_commands(message)


@bot.command(pass_context=True)
async def change_avatar(ctx, url: str):
    if ctx.message.author.id != '178887072864665600':
        await bot.say("You're not Zack enough.")
        return
    with aiohttp.ClientSession() as session:
        async with session.get(url) as resp:
            avatar_b = await resp.read()
            await bot.edit_profile(avatar=avatar_b)


@bot.command(pass_context=True)
async def dismiss(ctx):
    """Put the bot to sleep."""
    if ctx.message.author.id == '178887072864665600':
        await bot.say("Going to sleep.")
        await bot.logout()
    else:
        await bot.say("You wish.")
        return
