requirements.txt
discord.py
aiohttp
irc